// src/app/page/master/map-jenispegawai-jabatan/map-jenispegawai-jabatan.interface.ts
import { EventEmitter } from '@angular/core';

export interface MapJenispegawaiJabatan {
	kdProduk?: string;
	kdEvent?: number;
	label?: any;
	values?: any;
	value?: any;
	kode?: any;
	id?: any;
	kdProfile?: string;
	version?: string;
	reportDisplay?: string;
	kodeExternal?: string;
	namaExternal?: string;
	statusEnabled?: boolean;
	noRec?: string;
}

export interface DataSimpan {

	kdJenisPegawai?,
	mapJenisPegawaiToJabatanDto?,
}